apiserver/requirements.txt
amqp==5.2.0
asgiref==3.4.1
astroid==2.9.0
async-timeout==4.0.3
attrs==21.2.0
autopep8==1.6.0
billiard==4.2.0
black==21.11b1
celery==5.3.4
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-extensions==3.2.3
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.2.1
drf-spectacular==0.21.0
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.2.1
kombu==5.3.3
lazy-object-proxy==1.6.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.39
psycopg2==2.9.3
pycodestyle==2.9.0
PyJWT==2.3.0
pylint==2.12.2
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
PyYAML==6.0
redis==5.0.1
regex==2021.11.10
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.34.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
tomlkit==0.11.6
typing==3.7.4.3
typing_extensions==4.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.11
vine==5.1.0
wcwidth==0.2.9
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
